Made changes to all files
Made two turn classes for both players, it scans their hand and if it matches the requirements of the topcard (changes some variables in that class too), they are able to use the card.

Next step requires input from the user in which they say which card they want to play. Once done, it changes the value of the card and gives them a new card.