Turn system implemented for each player
Polished up existing methods, made it so players have a turn and can play a card. for loops are implemented so that they can iterate through the deck and see what cards the player has and whether or not they can play that card based on the top card of the pile.

Errors:
- When stating player 1's hand, it prints 14 cards instead of 7
- Top card changes each time it is mentioned through a print statement --> could be b/c the value changes inside of an if statement and the value doesn't get updated outside of it