Fixing errors
Fix made to 14 cards printing, now 7

Fix made to turn 1, where top card kept on changing between when topcard() was called and the restatement of the topcard in P1Turn() method. However, the topcard doesn't seem to change in P2Turn() method when called.

Error still in needing to update the players hand when they play the card and it is supposed to remove it from hand.